Enable F2FS
This allows both F2FS and EXT4 to be handled at runtime.
The intent is to prefer F2FS. But the bootloader chooses
what /data will be formated as when the device is flashed
for the 1st time.

Recovery wipes (reformats) /data with the default in fstab
 (in this case F2FS).

Uses the new
 TARGET_USERIMAGES_USE_F2FS
 BOARD_USERDATAIMAGE_FILE_SYSTEM_TYPE
 fstab_mgr that knows how to deal with entries for duplicate mount points.

Also, remove the un-used USE_E2FSPROGS.

Change-Id: I8d141a0d4d14df9fe84d3b131484e9696fcd8870
Signed-off-by: JP Abgrall <jpa@google.com>