BoardConfig+device: Don't build a userdata.img, don't use F2FS on non-linux
A userdata image is not very usefull as it does not match
the size of the device(s).
Don't waste time building it.

Also, on non-linux HOST_OS, don't bother with host F2FS tools.

The end result, is that on Mac OS, it will build volantis,
then fastboot will only erase userdata instead of formatting it,
and once it boots up it will format the erased paritition.


Change-Id: I7f2f6c92ca2af2da3fd741ecaab5d089c279c32b
Signed-off-by: JP Abgrall <jpa@google.com>