sound trigger: buffer streaming data from DSP
Adds a 16K buffer in the sound trigger for pulling streaming samples
from. This helps in two ways:

1. The number of trips through the IOCTL to the DSP is reduced, since we
can pull larger chunks of data down than whatever the audio HAL asks for
in a single shot. (There's a pretty substantial overhead in setting up
the SPI transfer in the kernel, so the fewer bursts we have to do the
better).

2. If capturing has been requested, we can now start pulling data from
the DSP while the upper layers are doing their own thing. It takes
roughly 30ms from the time the event is triggered until the audio HAL gets
around to asking for data, which is roughly how long it takes to pull
16K of data through the IOCTL.

Bug: 16729130
Change-Id: Ia48b3b5a69d1cce1dc973d714fbfff6ec5f8dc58