Update patch set 6
Patch Set 6: Code-Review-1

As for policy, we will be following 5.0 CDD - meaning all new devices (eg Ships with L as first release) should have encryption enabled by default. Section 9.9:

"For devices supporting full-disk encryption, the full-disk encryption SHOULD be enabled all the time
after the user has completed the out-of-box experience. While this requirement is stated as SHOULD for this version of
the Android platform, it is very strongly RECOMMENDED as we expect this to change to MUST in the future versions of
Android"

http://static.googleusercontent.com/media/source.android.com/en/us/compatibility/android-cdd.pdf

Patch-set: 6