Update patch set 6
Patch Set 6:

By booting a ROM with forceencrypt, the device will be encrypted, even if the user disabled forced encryption in the past (One way is to boot custom kernel after formatting all of userdata). On the other hand, booting a ROM with encryptable will NOT change the current encryption status, If a user flashes CyanogenMod, with stock settings (encryption on), the device will remain encrypted unless the user chooses to format userdata. Even if they choose to format userdata, they will have a choice in settings to use encryption or not.

Patch-set: 6