Update patch set 6
Patch Set 6: Code-Review+1

I have seen many users become mad with forceencrypt
especially me being a custom Kernel developer.

I think its vary bad to have it force encrypt
in security's name. Yes true having encryption
is vary good it adds security etc...
I understand that, but what i don't understand is why do we have to force people to use it why? 

In that thought its like your agreeing with locked bootloaders!
Its adding security, so lets force it, like the normal user wont even know anyway.

Sam Mortimer, Abhisek Devkota I agree with you
but CM is a custom ROM not a stock ROM.

To flash a ROM you need a unlocked bootloader
and a custom recovery.
now tell me that a unlocked bootloader is not a 
security risk and or a custom recovery

I'm just trying to tell you that a user accepts the risk
for unlocking the bootloader and/or flashing a ROM/recovery.

Note: 
having the kernel force encrypt that will use the stock password. 
A normal user wont even use a pin.
That in it self is a big security risk.
like all you have to do is flash a recovery that supports 
the stock password then mount data and BOOM security is compromised!

Patch-set: 6