Update patch set 6
Patch Set 6: Code-Review+1

If ROOT is offered to the users, we should give users this option as well; the security implications for root are just as large if not larger. In addition, the CDD makes this feature an option and being more restrictive than the CDD should be further justified.

Patch-set: 6