Update patch set 6
Patch Set 6:

Fair enough, but I ask you to reconsider, there is a reason why ROOT is still on CM and I think the same logic applies here. This is ROM that is empowering advanced users, undermining that value deserves more justification than people may be foolish and unlucky, there is a lot of ways CM gives you enough rope to hang yourself with.

Patch-set: 6