Update patch set 6
Patch Set 6:

"Fair enough, but I ask you to reconsider, there is a reason why ROOT is still on CM and I think the same logic applies here. This is ROM that is empowering advanced users, undermining that value deserves more justification than people may be foolish and unlucky, there is a lot of ways CM gives you enough rope to hang yourself with."

Yes, root is largely there "because it has been in the past". The justification is purely legacy. I fully consider root an exploit that we knowingly ship, and I would like to move away from that model (re: have been pushing for this for 4 years now). If there is functionality desired that root enables, I'd rather see a safe API built to handle it. 

CM as a _platform_ does not need to ship with root by default. There are various other options for users who wish to root, including such projects like SuperSu, or auto-root in recovery. It needn't be a platform default. Other projects have already adopted this model. 

Further, I would like to see us move away from test_key signed releases - which offers no serious verification of a proper CM authority (md5's on the download portal are download verification, not necessarily integrity verifications). 

Largely, I hate that we do things because that's the way its been.

Neither of these actions would _prevent_ the continued use of CM as a poweruser or developer base. Users who want such options have a recourse to get them, or compile on their own. But as the 'upstream' for a lot of other community projects, I feel our stance is frankly not beneficial to the betterment of the community.

Patch-set: 6