Update patch set 6
Patch Set 6:

Thank you for taking the time to express the justification for the decision, your position is internally consistent and a more than reasonable, although I would argue that ROOT is there for more than legacy reasons. 

I will only add that if the device is not encrypted it implies the user has taken some action to avoid it (or has formated their partition without rebooting), and since just booting the ROM causes an irreversible change to the system (or at least until you format it again), you are in most cases overruling the user without warning by encrypting the device on boot, and maybe this behavior should be enforced elsewhere, preferably with some sort of dialog where the user may be able to avoid the change.

Patch-set: 6