Enable dm-verity on flounder
Enable dm-verity for system and vendor partitions. Store the state in
the crypto metadata partition, and display a warning if dm-verity is
started in logging mode.

Adjusting read-ahead and prefetch_cluster settings for the dm-verity
devices reduces overhead to ~1% for sequential reads with block sizes
from 1kB to 32kB, and to ~9% for 512 byte blocks, when compared to the
best configuration for the raw block device.

Change-Id: Ia01ebab787ed65f52c7fb28abd4e0e2814bce97a