Update patch set 1
Patch Set 1:

Alright, understood.  I use my Nexus 9 for reading at night and even with live display the backlight is too bright for a dark room.  I can keep this stashed, but would like to share it with others as the low-end of the backlight brightness is not low enough. 

thanks

Patch-set: 1