allow mediacodec to read /system/lib dir
Very strange that this process needs to read the directory. It appears
to be listing the contents of the directory in order to find a library
name that contains "egl." Instead of the more typical method of opening
a known file name - in this case libegl.so.

Addresses:
avc: denied { read } for name="egl" dev="dm-1" ino=131
scontext=u:r:mediacodec:s0 tcontext=u:object_r:system_file:s0 tclass=dir

Bug: 27530325
Change-Id: I5dc91b912e9b7186ad6996df2e587f6ea242d896