init.recovery.flounder.rc: increase VM low water mark
system_server increases the low water mark by 36 MB during normal
Android userspace.  Match that in recovery mode.

volantis has a bad combination of large OTAs leading to a huge number of
active-file pages in flight, and an IOMMU driver that does page-sized
allocations in an atomic context.  This can lead to brief visual
glitches if the failed allocations cause us to drop a display flip.
Increasing the low-water mark greatly decreases the chances of this
happening.

Bug: 28025945
Bug: 28279077

Change-Id: I5a2d6a56b923ba3f3b45fbbcaf7d18f226b29c74
Signed-off-by: Greg Hackmann <ghackmann@google.com>