Update patch set 1
Patch Set 1:

(1 comment)

This leaves us with an unused flag (TARGET_USES_GRALLOC1) and the one that enables HWC2 (TARGET_USES_HWC2). According to Google, HWC2 shouldn't be enabled on flounder because it degrades performance. We can give it a go, but I would say to disable it in a later commit.

Patch-set: 1