Update patch set 17
Patch Set 17:

If several modifications are in fact needed to vendor, then yes I think it's probably best to build vendor img inline. And other custom roms have been doing it a while already on nougat, so it won't be too much of a shock to users.
One thing that comes to mind that we'd need to address, currently the proprietary files are unified for wifi and lte models. We'd probably want to separate them if we build vendor inline because of all the extra ril related stuff.  And the vendor img for lte model has about 50mb of radio firmware, that we'd have to include.

Patch-set: 17
Reviewer: razorloves <7118@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0