Update patch set 8
Patch Set 8:

Only negative thing I can think of at the moment is a user restoring a stock system backup and having it not work. If the user restores a system and vendor backup of course it will be fine. Even with this concern, I still believe that this is a better solution than moving everything to system, as using a new vendor image will consume less space on the system partition.

Patch-set: 8
Reviewer: Hashbang173 <13274@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0